requirements/base.txt
# Django Related
djangorestframework==3.15.2
Django==5.0.7
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-auditlog==3.0.0
drf-yasg==1.21.7
python-decouple==3.8


# Code Formatting
isort==5.13.2


#Database Related
psycopg2-binary==2.9.9
